Refactor the internal folder mess
