+++ src/internals/accounts/scanAccountsOnDevice.js
