+++ src/internals/accounts/sync.js
@@ -1,7 +1,7 @@
 // @flow
 
-export default (send: Function) => ({
-  all: () => {
-    setTimeout(() => send('accounts.sync.success'), 5e3)
-  },
-})
+import type { IPCSend } from 'types/electron'
+
+export default (send: IPCSend) => {
+  setTimeout(() => send('accounts.sync.success'), 5e3)
+}
