+++ src/internals/devices/index.js
@@ -1,0 +1,2 @@
+export listen from './listen'
+export checkIfAppOpened from './checkIfAppOpened'
