+++ src/internals/manager/index.js
@@ -1,0 +1,22 @@
+// @flow
+
+/**
+ *                                  Manager
+ *                                  -------
+ *
+ *                                    xXx
+ *                                    xXx
+ *                                    xXx
+ *                                  xxxXxxx
+ *                                   xxXxx
+ *                                    xXx
+ *                               xX    x    Xx
+ *                               xX         Xx
+ *                                xxXXXXXXXxx
+ *
+ */
+
+export { default as getMemInfos } from './getMemInfos'
+export { default as installApp } from './installApp'
+export { default as listApps } from './listApps'
+export { default as uninstallApp } from './uninstallApp'
