+++ src/internals/manager/installApp.js
@@ -1,0 +1,12 @@
+// @flow
+
+import type { IPCSend } from 'types/electron'
+
+import { createTransportHandler, installApp } from './helpers'
+
+export default (send: IPCSend, data: any) =>
+  createTransportHandler(send, {
+    action: installApp,
+    successResponse: 'manager.appInstalled',
+    errorResponse: 'manager.appInstallError',
+  })(data)
