+++ src/internals/manager/uninstallApp.js
@@ -1,0 +1,12 @@
+// @flow
+
+import type { IPCSend } from 'types/electron'
+
+import { createTransportHandler, uninstallApp } from './helpers'
+
+export default (send: IPCSend, data: any) =>
+  createTransportHandler(send, {
+    action: uninstallApp,
+    successResponse: 'manager.appUninstalled',
+    errorResponse: 'manager.appUninstallError',
+  })(data)
