--- src/internals/usb/index.js
@@ -1,3 +1,0 @@
-export devices from './devices'
-export wallet from './wallet'
-export manager from './manager'
